refactor: django auth-react test app
